% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_regression_wrappers.R
\name{basic_linear_regression}
\alias{basic_linear_regression}
\title{Performs basic linear regression with numerous Y vals}
\usage{
basic_linear_regression(expr_mat, p_data, model_formula)
}
\arguments{
\item{model_formula}{}
}
\description{
Performs basic linear regression with numerous Y vals
}
