% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_regression_wrappers.R
\name{basic_logistic_regression}
\alias{basic_logistic_regression}
\title{Performs logistic regression with numerous Y vals}
\usage{
basic_logistic_regression(expr_mat, p_data, model_formula)
}
\arguments{
\item{model_formula}{}
}
\description{
Performs logistic regression with numerous Y vals
}
