% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_wrappers.R
\name{enrich_go_wrapper}
\alias{enrich_go_wrapper}
\title{Performs enrichment analysis accross all GO domains}
\usage{
enrich_go_wrapper(ent_genes)
}
\arguments{
\item{gene_vector}{}
}
\description{
Performs enrichment analysis accross all GO domains
}
