% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_wrappers.R
\name{enrich_pathway_wrapper}
\alias{enrich_pathway_wrapper}
\title{Performs enrichment analysis using KEGG, REACTOME, DO, and DisGenNet}
\usage{
enrich_pathway_wrapper(ent_genes)
}
\arguments{
\item{ent_genes}{vector of entrez genes}
}
\value{
pathway list by database
}
\description{
Performs enrichment analysis using KEGG, REACTOME, DO, and DisGenNet
}
\keyword{internal}
