% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_wrappers.R
\name{ensembl_to_entrez_genes}
\alias{ensembl_to_entrez_genes}
\title{Title}
\usage{
ensembl_to_entrez_genes(gene_vector)
}
\arguments{
\item{gene_vector}{vector of human ensembl genes}
}
\value{
data.frame of with ensembl gene and entrez gene ids
}
\description{
Title
}
