% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{get_limma_results}
\alias{get_limma_results}
\title{get results from limma fit object}
\usage{
get_limma_results(limma_fit, coefs = NULL)
}
\arguments{
\item{limma_fit}{fit object produced by limma}

\item{coefs}{coefficients to return}
}
\value{
data.frame of results in long(ish) format
}
\description{
Takes a limma fit object and returns a long data.frame
of the results. If no coefficients are passed, it tries
to guess which are the relevant ones checking which
are made up of 1's and 0's.
}
