% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_categorical_triplot.R
\name{pca_categorical_triplot}
\alias{pca_categorical_triplot}
\title{Title}
\usage{
pca_categorical_triplot(plot_data, ax_labels, plot_ellipse = F,
  plot_title = "", plot_legend = TRUE)
}
\arguments{
\item{plot_legend}{}
}
\description{
Title
}
