% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_functions_main.R
\name{pca_plotter}
\alias{pca_plotter}
\title{Plots PC1 by PC2 using ggplot}
\usage{
pca_plotter(transformed_data, sample_map, leg_row_num = 3,
  gene_num = Inf, return_data = FALSE)
}
\arguments{
\item{return_data}{}
}
\description{
Plots PC1 by PC2 using ggplot
}
