% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{uniquefy_by_abs_max}
\alias{uniquefy_by_abs_max}
\title{Return a unique data.frame by max value}
\usage{
uniquefy_by_abs_max(mat, group_col = NULL, val_col = NULL)
}
\arguments{
\item{mat}{data.frame with 1st column as string and second as numeric}
}
\value{
unique data.frame
}
\description{
Return a unique data.frame by max value
}
\examples{
set.seed(42)
mat <- data.frame(group=c('a', 'a', 'a', 'b', 'c', 'c'), vals=runif(6))
mat$vals[6] <- -4
}
