% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{uniquefy_by_variance}
\alias{uniquefy_by_variance}
\title{Title}
\usage{
uniquefy_by_variance(expr_mat, probe_set)
}
\arguments{
\item{probe_set}{}
}
\description{
Idea for using max var to de-duplicate probes came
form here: https://www.biostars.org/p/51756/#51875
Probe set needs to be a data frame with one column
the rownames are the probe ID and the single column is
refseq or equivalent gene annotation
}
