% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{volcano_plotter}
\alias{volcano_plotter}
\title{Creates a volcano plot using data in long format}
\usage{
volcano_plotter(plot_data, plot_labels = TRUE, label_num = 10)
}
\arguments{
\item{plot_data}{plot data in long format}

\item{plot_labels}{whether labels should be drawn for some genes}

\item{label_num}{number of genes to draw labels}
}
\value{
ggplot object
}
\description{
This function creates a volc plot using the returned results
in long format. Meant to be used with my results wappers
}
