% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_categorical_triplot.R
\name{pca_categorical_triplot}
\alias{pca_categorical_triplot}
\title{PCA categorical triplot}
\usage{
pca_categorical_triplot(plot_data, ax_labels, plot_ellipse = F,
  plot_title = "", plot_legend = TRUE)
}
\arguments{
\item{plot_legend}{}
}
\description{
Currently this is too unweildy to use out in the wild
but it's not a bad overall, just needs some work.
}
\keyword{internal}
